// portal/src/_variables.scss.d.ts
export interface I_globalScss {
    primaryLight1: string;
    primaryLight2: string;
    secondaryLight1: string;
    secondaryLight2: string;
    tertiaryLight1: string;
    background: string;
    accentLight: string;
    accentDark: string;
    accentDarker: string;
    error: string;
    fontNormal: string;
    fontSemibold: string;
    fontBold: string;
}

export const styles: I_globalScss;

export default styles;